m8: Import Android 6.0 GPE source 5.07.1700.6
* m8ulstockui-3.4.0-gd7aafa47

Conflicts:
	AndroidKernel.mk
	arch/arm/boot/dts/dsi-panel-mem.dtsi
	arch/arm/common/gic.c
	arch/arm/crypto/aes-armv4.S
	arch/arm/crypto/sha1-armv4-large.S
	arch/arm/mach-msm/Makefile
	arch/arm/mach-msm/acpuclock-8974.c
	arch/arm/mach-msm/acpuclock-cortex.c
	arch/arm/mach-msm/bam_dmux.c
	arch/arm/mach-msm/board-8226-audio.c
	arch/arm/mach-msm/board-htc-8974.c
	arch/arm/mach-msm/board-m8-audio.c
	arch/arm/mach-msm/board-m8-gpiomux.c
	arch/arm/mach-msm/clock-8226.c
	arch/arm/mach-msm/clock-8974.c
	arch/arm/mach-msm/clock-a7.c
	arch/arm/mach-msm/clock-generic.c
	arch/arm/mach-msm/clock-krait-8974.c
	arch/arm/mach-msm/clock-krait.c
	arch/arm/mach-msm/clock-pll.c
	arch/arm/mach-msm/cpr-regulator.c
	arch/arm/mach-msm/display/htc_a5-panel.c
	arch/arm/mach-msm/htc_footprint.c
	arch/arm/mach-msm/htc_mnemosyne.c
	arch/arm/mach-msm/htc_monitor.c
	arch/arm/mach-msm/htc_restart_handler.c
	arch/arm/mach-msm/htc_smem.h
	arch/arm/mach-msm/include/mach/cable_detect.h
	arch/arm/mach-msm/include/mach/htc_footprint.h
	arch/arm/mach-msm/include/mach/msm_rtb.h
	arch/arm/mach-msm/msm-pm.c
	arch/arm/mach-msm/msm_qmi_interface.c
	arch/arm/mach-msm/msm_rtb.c
	arch/arm/mach-msm/msm_watchdog_v2.c
	arch/arm/mach-msm/qdsp6v2/voice_svc.c
	arch/arm/mach-msm/restart.c
	arch/arm/mach-msm/rpm_htc_cmd_shared.h
	arch/arm/mach-msm/subsystem_restart.c
	drivers/block/zram/zram_drv.c
	drivers/char/hw_random/msm_rng.c
	drivers/cpufreq/cpufreq.c
	drivers/crypto/msm/qcrypto.c
	drivers/gpu/ion/ion_heap.c
	drivers/gpu/msm/adreno_drawctxt.c
	drivers/gpu/msm/adreno_drawctxt.h
	drivers/gpu/msm/adreno_ringbuffer.c
	drivers/gpu/msm/adreno_snapshot.c
	drivers/gpu/msm/kgsl_sync.c
	drivers/i2c/chips/rt5501.c
	drivers/input/touchscreen/himax_852xD.c
	drivers/input/touchscreen/synaptics_3k.c
	drivers/media/platform/msm/camera_v2/isp/msm_isp_util.c
	drivers/media/platform/msm/camera_v2/sensor/actuator/msm_actuator.c
	drivers/media/platform/msm/camera_v2/sensor/msm_sensor.c
	drivers/media/platform/msm/vidc/msm_smem.c
	drivers/media/platform/msm/vidc/msm_vdec.c
	drivers/media/platform/msm/vidc/msm_venc.c
	drivers/media/platform/msm/vidc/msm_vidc.c
	drivers/media/platform/msm/vidc/msm_vidc_common.c
	drivers/media/platform/msm/vidc/msm_vidc_common.h
	drivers/media/platform/msm/vidc/msm_vidc_internal.h
	drivers/media/platform/msm/vidc/venus_hfi.c
	drivers/media/radio/radio-iris-transport.c
	drivers/misc/Makefile
	drivers/misc/qpnp-vibrator.c
	drivers/misc/qseecom.c
	drivers/misc/uid_cputime.c
	drivers/mmc/Kconfig
	drivers/mmc/card/block.c
	drivers/mmc/card/queue.c
	drivers/mmc/core/cd-gpio.c
	drivers/mmc/core/core.c
	drivers/mmc/core/host.c
	drivers/mmc/core/mmc.c
	drivers/mmc/core/sd.c
	drivers/mmc/host/sdhci-msm.c
	drivers/mmc/host/sdhci.c
	drivers/net/kineto_gan.c
	drivers/net/wireless/wcnss/wcnss_vreg.c
	drivers/net/wireless/wcnss/wcnss_wlan.c
	drivers/platform/msm/avtimer.c
	drivers/rtc/alarm.c
	drivers/slimbus/slim-msm-ngd.c
	drivers/slimbus/slim-msm.c
	drivers/spmi/spmi-dbgfs.c
	drivers/staging/zram/zram_drv.h
	drivers/thermal/msm_thermal.c
	drivers/usb/dwc3/gadget.c
	drivers/usb/gadget/ci13xxx_udc.c
	drivers/usb/gadget/f_mtp.c
	drivers/usb/gadget/u_xpst.c
	drivers/usb/otg/msm_otg.c
	drivers/video/msm/mdss/mdss_dsi_panel.c
	drivers/video/msm/mdss/sii8240_8558/mhl_supp.c
	drivers/video/msm/mdss/sii8240_8558/si_mhl_defs.h
	drivers/video/msm/mdss/sii8240_8558/si_mhl_tx_hw_drv_api.h
	fs/block_dev.c
	fs/exec.c
	fs/ext4/file.c
	fs/ext4/inode.c
	fs/ext4/mballoc.c
	fs/fat/inode.c
	fs/fs_struct.c
	fs/inode.c
	fs/jbd2/journal.c
	fs/mount.h
	fs/namespace.c
	fs/open.c
	fs/pnode.h
	fs/proc/base.c
	fs/proc/generic.c
	fs/proc/namespaces.c
	fs/sync.c
	include/linux/CwMcuSensor.h
	include/linux/cm3629.h
	include/linux/cpufreq.h
	include/linux/fs.h
	include/linux/lsm_audit.h
	include/linux/mm_types.h
	include/linux/mmc/card.h
	include/linux/mmc/host.h
	include/linux/mmc/mmc.h
	include/linux/mmc/sdhci.h
	include/linux/pid_namespace.h
	include/linux/prctl.h
	include/linux/proc_fs.h
	include/linux/videodev2.h
	include/linux/wcnss_wlan.h
	include/media/msm_cam_sensor.h
	include/media/radio-iris.h
	include/net/addrconf.h
	include/net/netlink.h
	include/net/tcp.h
	include/sound/apr_audio-v2.h
	include/sound/q6asm-v2.h
	include/sound/soc.h
	include/trace/events/mmcio.h
	kernel/async.c
	kernel/power/htc_pnpmgr.c
	kernel/power/power.h
	kernel/power/process.c
	kernel/sched/core.c
	kernel/sys.c
	kernel/trace/trace.c
	mm/debug-pagealloc.c
	mm/ksm.c
	mm/page_alloc.c
	mm/slub.c
	mm/vmalloc.c
	mm/vmscan.c
	mm/zsmalloc.c
	net/ipv4/tcp.c
	net/ipv4/udp.c
	net/ipv6/addrconf.c
	net/ipv6/udp.c
	scripts/gcc-wrapper.py
	security/selinux/avc.c
	security/selinux/hooks.c
	security/selinux/include/objsec.h
	security/selinux/ss/policydb.c
	security/selinux/ss/services.c
	sound/core/pcm.c
	sound/core/pcm_lib.c
	sound/core/pcm_native.c
	sound/soc/codecs/wcd9306.c
	sound/soc/codecs/wcd9320.c
	sound/soc/codecs/wcd9xxx-mbhc.c
	sound/soc/codecs/wcdcal-hwdep.c
	sound/soc/msm/qdsp6v2/msm-compr-q6-v2.c
	sound/soc/msm/qdsp6v2/msm-compress-q6-v2.c
	sound/soc/msm/qdsp6v2/msm-pcm-q6-v2.c
	sound/soc/msm/qdsp6v2/msm-pcm-q6-v2.h
	sound/soc/msm/qdsp6v2/msm-pcm-voip-v2.c
	sound/soc/msm/qdsp6v2/q6afe.c
	sound/soc/msm/qdsp6v2/q6asm.c
	sound/soc/msm/qdsp6v2/q6core.c
	sound/soc/soc-compress.c
	sound/soc/soc-pcm.c

+++ block/row-iosched.c
@@ -165,7 +165,7 @@ static inline void __maybe_unused row_dump_reg_and_low_stat(struct row_data *rd)
 	for (i = ROWQ_PRIO_REG_SWRITE; i < ROWQ_MAX_PRIO; i++) {
 		if (!list_empty(&rd->row_queues[i].fifo)) {
 			struct request *check_req = list_entry_rq(rd->row_queues[i].fifo.next);
-			unsigned int check_jiffies = ((unsigned long) (check_req)->csd.list.next);
+			unsigned long check_jiffies = ((unsigned long) (check_req)->csd.list.next);
 
 			if (time_after(jiffies,
 			    check_jiffies + msecs_to_jiffies(ROW_DUMP_REQ_STAT_MSECS))) {
